tests/submission/Microsoft/audio/audio_004.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'HTMLAudioElement' inherits 'NETWORK_NO_SOURCE' from 'HTMLMediaElement' and supports getting its value</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-network_no_source" />
  <meta name="assert" content="'HTMLAudioElement' inherits 'NETWORK_NO_SOURCE' from 'HTMLMediaElement' and supports getting its value" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>

    <div id=log></div>
    <audio id="audio0">Your browser does not support media elements.</audio>
    <script type="text/javascript">
        test(function() {var testElement = document.getElementById("audio0"); assert_equals(testElement.NETWORK_NO_SOURCE, 3)}, "'HTMLAudioElement' inherits 'NETWORK_NO_SOURCE' from 'HTMLMediaElement' and supports getting its value");
        test(function() {var testElement = document.getElementById("audio0"); testElement.NETWORK_NO_SOURCE = 10; assert_equals(testElement.NETWORK_NO_SOURCE, 3)}, "The inherited 'NETWORK_NO_SOURCE' property of the 'HTMLAudioElement' interface is a constant");
    </script>
</body>
</html>
